backend/pyproject.toml
[project]
name = "langchain-learning-backend"
version = "0.1.0"
description = "LangChain Learning Project - Backend"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "deepagents>=0.1.0",
    "dashscope>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "sse-starlette>=2.1.0",
    "python-multipart>=0.0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "common", "demos", "models"]
